Revert "adapted to moved AbstractWorkflowBasedRunConfiguration"
This reverts commit 89a3a37ea4dd44dd1c3f8d26c10522fe67c1e4f3.
